disable logger, fix

+++ quant.py
@@ -178,7 +178,10 @@ class Quant4Linear(nn.Module):
         dtype = x.dtype
         if self.faster:
             x = x.half()
-            quant_cuda.vecquant4matmul_faster(x, self.qweight, y, self.scales, self.zeros)
+            y = y.half()
+            self.scales = self.scales.half()
+            self.zeros = self.zeros.half()
+            quant_cuda.vecquant4matmul_faster(x, self.qweight, y, self.scales.half(), self.zeros.half())
         else:
             x = x.float()
             quant_cuda.vecquant4matmul(x, self.qweight, y, self.scales, self.zeros)
